+++ src/lib/services/types/verification/wasm.ts
@@ -1,0 +1,22 @@
+import { z } from "zod";
+
+import { zWasmVerifyInfo } from "lib/types";
+
+export interface WasmVerifyRequest {
+  chainId: string;
+  codeId: number;
+  gitUrl: string;
+  commit: string;
+  packageName: string;
+  compilerVersion: string;
+}
+
+export const zWasmVerifyInfosResponse = z.record(
+  z.coerce.number(),
+  zWasmVerifyInfo.nullable()
+);
+
+export const zWasmRelatedVerifyInfosResponse = z.record(
+  z.string(),
+  zWasmVerifyInfo.omit({ verificationInfo: true })
+);