+++ src/lib/utils/index.ts
@@ -38,5 +38,6 @@ export * from "./tx";
 export * from "./txHash";
 export * from "./validate";
 export * from "./validator";
+export * from "./verification";
 export * from "./window";
 export * from "./zod";